refactor(surface): put the blur-chain accessor where packs can reach it
All seven chain packs sampled iChannel6 by hand, seventeen times between
them, while surfaceBlurTexel sat unused. That looked like laziness and
was not: the helper lived in surface_blur.glsl, which no main pass
includes. The dependency runs the other way, surface_blur.glsl includes
surface_multipass.glsl, and a main pass includes only that one. Calling
the helper would have meant pulling the whole pyramid, its reach tables
and its combiners, into seven main passes for one line. Trying it first
is how this came out: every chain pack failed to bake.

So the helper moves to surface_multipass.glsl, beside the iChannel6
declaration it reads and in the header a main pass actually includes.
The chain passes still see it. Every pack goes through it now, which
matters because the channel the chain lands on is a contract between the
pass list and the main pass: seventeen copies of that slot number are
seventeen places to miss if it ever moves, and none of them would fail
to compile. They would read whatever else was bound.

The builtin gaussian pair is also baked by a test for the first time.
No bundled pack declares those tokens, so no gate ever compiled them,
and the other in-tree references only resolve the token and stop. A pack
outside this repo was the first thing to compile them. Breaking
gaussian_h.frag on purpose now fails exactly one slot and nothing else.

And setBufferScales had the failure mode backwards in its own comment. It
said qBound on a NaN is unspecified. Reading Qt's qminmax.h and running
it: qBound is qMax(min, qMin(max, val)), both comparisons against NaN are
false, and the answer is the minimum. So a quoted number and a NaN are
indistinguishable in the result, both landing on a hundred and
twenty-eighth of the canvas. That is a better argument for the guard than
the one the comment made.

+++ libs/phosphor-rendering/src/shadernoderhiuniforms.cpp
@@ -115,8 +115,13 @@ void ShaderNodeRhi::syncBaseUniforms(QRhi* rhi)
 
     // iChannelResolution — resolved node-side (needs live RHI textures).
     const bool multiBufferMode = m_bufferPaths.size() > 1;
-    const int numChannels = multiBufferMode ? qMin(m_bufferPaths.size(), static_cast<qsizetype>(kMaxBufferPasses))
-                                            : (m_bufferShaderReady && m_bufferTexture ? 1 : 0);
+    // The cast is explicit because the qMin answers a qsizetype and this is an
+    // int. It cannot lose anything, since the same qMin bounds it by
+    // kMaxBufferPasses, but an implicit narrowing here is indistinguishable at a
+    // glance from one that can.
+    const int numChannels = multiBufferMode
+        ? static_cast<int>(qMin(m_bufferPaths.size(), static_cast<qsizetype>(kMaxBufferPasses)))
+        : (m_bufferShaderReady && m_bufferTexture ? 1 : 0);
     for (int i = 0; i < PhosphorShaders::Bindings::kChannelResolutionSlots; ++i) {
         if (i < numChannels) {
             if (multiBufferMode && m_multiBufferTextures[i]) {
